requirements.txt
aiogram==3.4.1
openai==1.54.0
httpx==0.25.2
requests==2.31.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
fastapi==0.109.2
uvicorn==0.27.1
sqladmin==0.16.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytest==8.0.0
pytest-asyncio==0.23.5
hypothesis==6.98.15
aioresponses==0.7.6
pytest-mock==3.12.0
Pillow==10.2.0